# syllab-backend/openapi.yaml
openapi: 3.0.3
info:
  title: SylLab API
  description: |
    Anti-Plagiarism Code Forensics Platform.
    Establishes proctored Week 1 baselines and detects AI-assisted code submissions
    through AST-based sophistication scoring and cohort comparison.
  version: 1.0.0
  contact:
    name: SylLab Support
    email: support@syllab.io
servers:
  - url: https://api.syllab.io/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /sessions:
    post:
      tags: [sessions]
      summary: Create a proctored baseline session (Instructor/Proctor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              course_id: crs-101
              scheduled_at: '2025-09-01T09:00:00Z'
              duration_minutes: 30
              room_label: Lab A-204
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [sessions]
      summary: List sessions for a course
      parameters:
        - name: course_id
          in: query
          required: true
          schema:
            type: string
          example: crs-101
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /submissions:
    get:
      tags: [submissions]
      summary: List submissions for a student or course
      parameters:
        - name: student_id
          in: query
          schema:
            type: string
          example: usr-001
        - name: course_id
          in: query
          schema:
            type: string
          example: crs-101
        - name: week_number
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
          example: 4
        - name: flag_level
          in: query
          schema:
            type: string
            enum: [NORMAL, MONITOR, REVIEW, INTERVIEW]
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, -created_at, ensemble_score, -ensemble_score]
            default: -created_at
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated submissions list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── COHORTS ───────────────────────────────────────────────────────────────
...
  /instructor/queue:
    get:
      tags: [instructor]
      summary: Get prioritized action queue of flagged submissions
      parameters:
        - name: course_id
          in: query
          schema:
            type: string
          example: crs-101
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, REVIEWED, CLEARED, CONFIRMED]
            default: PENDING
        - name: min_score
          in: query
          schema:
            type: number
            format: float
          example: 1.5
        - name: sort
          in: query
          schema:
            type: string
            enum: [priority_score, -priority_score, created_at, -created_at]
            default: -priority_score
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated action queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /auth/login.
        Include as: Authorization: Bearer <token>
        Tokens expire after 24 hours.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: |
        Cursor for keyset pagination. Pass the value of `next_cursor`
        from the previous response. Omit for first page.
      example: eyJpZCI6InN1Yi0wNDIifQ
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
      example: 20
...
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [STUDENT, INSTRUCTOR, PROCTOR, ADMIN]

    # ── SESSIONS ────────────────────────────────────────────────────────────
    CreateSessionRequest:
      type: object
      required: [course_id, scheduled_at, duration_minutes]
      properties:
        course_id:
          type: string
        scheduled_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          minimum: 10
          maximum: 120
          example: 30
        room_label:
          type: string
          example: Lab A-204
    Session:
      type: object
      properties:
        id:
          type: string
          example: sess-001
        course_id:
          type: string
        status:
          type: string
          enum: [SCHEDULED, OPEN, CLOSED]
        scheduled_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        room_label:
          type: string
        opened_at:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        network_isolation_confirmed:
          type: boolean
        baseline_count:
          type: integer
          description: Number of students who submitted baseline in this session
    SessionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    SophisticationFeatures:
      type: object
      description: AST-extracted complexity metrics
      properties:
        total_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        error_handling_tier:
          type: integer
          enum: [0, 1, 3, 5]
          description: "0=none, 1=basic try/catch, 3=specific exceptions, 5=custom hierarchies"
        architecture_tier:
          type: integer
          enum: [0, 2, 4, 5]
          description: "0=flat functions, 2=simple classes, 4=service layer, 5=repository/CQRS"
        type_safety_score:
          type: integer
          enum: [0, 1, 3]
          description: "0=none, 1=partial hints, 3=comprehensive Pydantic/TypeScript"
        has_decorators:
          type: boolean
        has_async:
          type: boolean
        has_context_managers:
          type: boolean
        has_metaclasses:
          type: boolean
        naming_verbosity_score:
          type: number
          format: float
          description: Average identifier length normalized 0-1
        cyclomatic_complexity_avg:
          type: number
          format: float
        max_nesting_depth:
          type: integer
        has_dependency_injection:
          type: boolean
        has_abstract_classes:
          type: boolean
        has_dataclasses:
          type: boolean
        comment_density:
          type: number
          format: float
          description: Comments per line of code
    GenealogyViolation:
      type: object
      properties:
        technique:
          type: string
          example: circuit_breaker
        missing_prerequisite:
          type: string
          example: context_managers
        severity:
          type: number
          format: float
          example: 0.5
    AnalysisResult:
      type: object
      properties:
        submission_id:
          type: string
        flag_level:
          type: string
          enum: [NORMAL, MONITOR, REVIEW, INTERVIEW]
        ensemble_score:
          type: number
          format: float
        score_breakdown:
          type: object
          properties:
            trajectory_z:
              type: number
              format: float
            genealogy_penalty:
              type: number
              format: float
            cohort_outlier:
              type: number
              format: float
            weights:
              type: object
              properties:
                trajectory:
                  type: number
                  format: float
                genealogy:
                  type: number
                  format: float
                cohort:
                  type: number
                  format: float
        sophistication:
          $ref: '#/components/schemas/SophisticationFeatures'
        genealogy_violations:
          type: array
          items:
            $ref: '#/components/schemas/GenealogyViolation'
        cohort_position:
          type: object
          properties:
            week_number:
              type: integer
            percentile:
              type: number
              format: float
            cohort_mean:
              type: number
              format: float
            cohort_std:
              type: number
              format: float
        weeks_compressed:
          type: number
          format: float
          description: Equivalent weeks of learning detected in this single submission
        baseline_score:
          type: number
          format: float
        expected_score_week4:
          type: number
          format: float
    SubmissionReport:
      allOf:
        - $ref: '#/components/schemas/AnalysisResult'
        - type: object
          properties:
            submitted_at:
              type: string
              format: date-time
            course_id:
              type: string
            week_number:
              type: integer
            language:
              type: string
    SubmissionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionReport'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # ── COHORTS ─────────────────────────────────────────────────────────────
...
    QueueItem:
      type: object
      properties:
        id:
          type: string
        submission_id:
          type: string
        student:
          $ref: '#/components/schemas/UserSummary'
        priority_score:
          type: number
          format: float
        status:
          type: string
          enum: [PENDING, REVIEWED, CLEARED, CONFIRMED]
        llm_guidance:
          type: string
          nullable: true
          description: Gemini-generated interview guidance. Null if score < 2.5 or API call failed.
        instructor_note:
          type: string
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    QueueListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/QueueItem'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      description: Cursor-based pagination metadata
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Pass as `cursor` parameter in next request. Null if no more pages.
          example: eyJpZCI6InN1Yi0wNDIifQ
        has_more:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required: [error, message, status]
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: BASELINE_NOT_FOUND
        message:
          type: string
          description: Human-readable description
          example: No locked baseline found for student in course crs-101
        status:
          type: integer
          example: 404
        details:
          type: array
          description: Field-level validation errors (422 only)
          items:
            type: object
            properties:
              field:
                type: string
                example: week_number
              message:
                type: string
                example: Must be between 1 and 20
  responses:
    BadRequest:
      description: Invalid request format or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: BAD_REQUEST
            message: Invalid request body
            status: 400
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: JWT token is missing or expired
            status: 401
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Instructor role required to access this resource
            status: 403
...
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed
            status: 422
            details:
              - field: week_number
                message: Must be between 1 and 20
              - field: language
                message: Must be one of python, javascript, typescript
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INTERNAL_SERVER_ERROR
            message: An unexpected error occurred
            status: 500